Improved search by payer and improved feedback.
If any of the objects are searching, then they will have a spinner over
them.  Right now, the checkouts object is the exception.  We don't want
the spinner to appear everytime we update the table.  We only want it
when the table is being loaded for the first time, and even then that's
not perfect.